Created ImageLoader component, and split it out from CppPlugin
The CppPlugin had become a component and I didn’t like that, as
essentially it should just be a wrapper class around the underlying C++
code.

So I broke out part of the CppPlugin into a component called
ImageLoader, such that the CppPlugin is back to just being a normal
class, and the ImageLoader is the component that the other systems
interface with.

Currently the CppPlugin is used just for loading images, so its only
the ImageLoader that speaks to it, but in the future we might find that
the CppPlugin is actually used for other things, so other classes might
need access to it - so I think that in principle this split is tidier
and more correct for future proofing those situations.