Added partial Caching of Original Images
When Original Images are loaded we now cache them into the ImageSphere
that they belonged to, hence replacing the thumbnail that was stored
there.

I’m calling this only partial caching, because as soon as you switch
into the next set of spheres (by pressing one the gallery, previous or
next buttons), then you lose all of those Original Images.

I’ve spotted that although its great the thumbnail loading is very
fast, its pretty frustrating to click on a sphere and not get immediate
feedback. The fact that every time you click on a sphere you get a
loading icon seems very annoying - its like you are being punished for
being curious.

I think to improve it I need to make it so that the loading process is
fun - a good idea might be to show the image being loaded in the
background.