Added option to Delete a post
I added the option to delete an image on your profile.

I’ve put the Icon in a far too visible place, as the user will likely
press it just when playing around, but I’m not sure how best to place
it right now, and I had to do it because I’d accidentally added a post
twice and wanted to delete it!