Added capacity for app to keep you logged in between sessions!
I now write the “uid”, “client”, and “access-token” variables into a
tiny binary file on the user’s device that means I can keep users
logged in between sessions. Every time any of these values updates we
write the file again (this may not be the most efficient way of doing
it, but it guarantees correctness).

Then when user signs out the file is deleted so that the next time they
log into the app it will not be able to find their login details and
they will be presented with the login screen.

I had to create a new state in the AppDirector to have this work, and I
also removed its reliance on coroutineQueues, two things I’m not so
certain about, but we’ll see how it goes!