Fixed that I wasn't getting Handle, Email and Name on secondary logins
To get hold of the Handle, Email, Name and Profile I required to do a
SignIn, but since I’m not storing the user’s Username or Password I
couldn’t do that anymore!

But Dan added a new GET to ‘/users’ method to the API meaning that I
can now get hold of these values even on a secondary login (as in when
a user has logged in, closed the app, then re-opened it and they are
still signed on).

The Profile is therefore now correctly showing the welcome text!