Added 2-step Confirm/Back options for deleting posts
To delete a post you now have to click through a Confirm button to
ensure you don’t do it by accident, and I’ve also added text explaining
what you are up to in the message I have in the centre of the UI.

In the process of doing this I also created prefabs for two Text Button
types (Large chunks of text, and small quick pieces of text), replacing
the old way of writing things into a PSD file and exporting it.

I also put simply one piece of User Text in the centre (as before it
was split between Gallery and Profile text)!