Added support for Setting Password, and Deleting Account, plus fixed a bug with PasswordText code
The OptionsFlow now has implementations for Setting Password and
Deleting Account, which essentially just call through to BackEndAPI’s
Register_UpdateUser() and Register_DeleteUser() respectively!

The small PasswordText component I’d written had a silly bug meaning it
wasn’t checking the correct variable, and hence I was sometimes sending
the wrong password over to the API.

I also moved the Keyboard closer to the camera because it was
intercepting with the Options text!