Performance fixes! In all cases I ended up just moving code off the main thread.
[BackEndAPI] Made Deserialisation cheaper by simply threading it out!
As it seemed like Deserialisation was surprisingly expensive.

[User] Similarly, I found that writing my little binary file out every
time an API call updated the AccessToken was also expensive, so I
threaded that out too.

[ImageLoader] Once again, I found that both (1) doing the
WebRequest.Create(url) call and consequently (2) getting the
imageStream out of that request were causing slowdowns, so I bunched
them into a function and threaded that all off too!