Perfomance!
Currently I’ve got the Analytics working again.

And I’ve moved DeviceGallery’s OpenAndroidGallery() +
ImageSphereController’s SetAllImageSpheresToLoading() and
HideAllImageSpheres() to work as Coroutines.

As well as replacing all “yield return new WaitForEndOfFrame();” for
“yield return null” in order to reduce GC problems.

Plus I stopped “Resources.UnloadUnusedAssets();” being called
practically every frame from FPSSystem and put it only every so often
in AppDirector (when we press a button on the MenuBar).

However, all of this has actually created more spikes that seem to be
manifesting in CppPlugin - so I’m not in anyway close to checking this
in…