Made it such that Uploading Images is now threaded!
By threading each part of the UploadImageInternal() function, we now
have that entire functionality running in frame. =D

The main thing that I found very beneficial was that I by adding
AndroidJNI.AttachCurrentThread() to the start, and
AndroidJNI.DetachCurrentThread() to the end of my Java function, I was
able to thread it! I will now try using it for my other Java plugin
calls!