Upped Thumbnail size, hid keyboard correctly, fixed double click bug
Thumbnails are now 640px’s width, so double the size I had them at
before - this makes the image a fair bit clearer and has no notable
impact on performance.

The keyboard is now also appearing and hiding correctly when the user
slides their finger up or down.

And I fixed an annoying bug where MenuButton would always call the
HandleUp() function twice due to the fact that I have some functions in
there to make it possible for you to call the code correctly even when
you are not in VR mode. This meant that MouseUp() was calling through
to HandleUp() as well as just the normal VR Interactive Button method
of doing things.