Caption upload and UI added + Made Keyboard conserve Text
This simply follows the same format as deleting images (with a confirm
or cancel option), plus it has a text box for the user to write the
caption into.

Previously, every time you clicked on a text box it would remove the
text that was there, but I actually find this pretty annoying so I’ve
changed it to conserve the text unless its a password (and the password
thing is more to do with asterixes being a pain)

I’m also grabbing captions from the Profile side of things, but I
haven’t got a way to display them yet - the next thing will be to
display the caption.