Added Handle, Likes, and Captions to each ImageSphere in the UI
Now ImageSphere’s will be given this “metadata” when the image is set,
and be able to display it.

This required quite a significant shuffle around of the UI, but
essentially was just a job of pushing out the spheres so that they have
more space around them for each of the Handle, Likes and Captions.

I’ve also moved over to using the “vreel-development” environment but
somehow this has led to every API call returning “Authorised
Applications Only” - I’m checking in this current work in but need to
work out what is going wrong with the API =(