Added the Search SubMenu + Small tweak to Captions + White version of all Icons
The Search SubMenu now exists although its still entirely empty.

Captions will be hidden so that they do not change in size with
ImageSpheres as they come in and out.

The button images all have white versions now as I thought I might turn
the buttons into something similar to my black buttons - however, as
soon as I tested that, it led to massive amounts of scintillation,
which is just not worth it.