Created Search class + Renamed DeviceGallery to Gallery
We now have a Search class that interacts with the AppDirector just the
same way that Profile and DeviceGallery do, although currently its a
blank class that simply just hides the Search Results.

I initially named the class DeviceGallery but I have always internally
called it simply Gallery, therefore I’ve renamed it to align with my
interpretation of the class. Now we have Profile, Search and Gallery,
in much the same way that we have the buttons to represent them.