Autocomplete User and HashTag searching now works!
The way I see it Searching has 2 steps:
1) Searching for Users and Hashtags
2) Selecting Users and Hashtag and getting results

I’ve managed to get step 1 complete such that now as you type into the
search box it will display the first 10 users/hashtag results that the
system can find! =D

The next thing to do here will be to create the option of selecting
these results and viewing the appropriate user or posts associated with
a tag!