Added Selecting a User I've searched for!
We can now search for a User and then select that result and view their
profile! This is the first real feeling of the app being social!

The way this works is that the Search system delegates the work over to
the Profile, and hence there’s not that much new code that’s actually
been written, and all that’s changed inside the Profile is that its now
aware of the UserID its performing its search for!

Since everything makes use of Next and Previous arrows, I’ve moved them
into the MenuBar and have them discover whether they should be visible
or not depending on the current state of the AppDirector - before they
used to be within their relevant SubMenu’s eg. ProfileSubMenu and
GallerySubMenu.