Got HashTag searching and selection to work!
In principle all of searching is now working! I’m just going to run
through it tomorrow to see if I spot any bugs before merging into
Master!

In order to do HashTag Search I realised that I would be making use of
the same code as Profile (in much the same way as User Search was),
therefore I broke out the re-usable structure of Profile into a new
Posts component!!

This Posts component essentially knows whether its dealing with a
UserProfile, OtherProfile or a HashTag result, and therefore calls the
appropriate function in order to get hold of all the posts data, but
the component goes onto handle this data all the same regardless.

The simplicity and elegance of this solution means I’ll likely be able
to re-use it for the Timeline as well! =)