Correctly fishing out user Handle for Posts found through Search
Using the ID from “relationships” I need to iterate over the “Included”
array in Model_Posts and find the user to find his handle.

Atm this “included” array only has UserData, but in reality it could
hold anything, in the long term I’ll need to fix it such that I check
the “type” field and deserialise based on that.

This now means that for every post you get the handle, likes and
caption!