Repositioned "People" and "Tags" buttons
These were clashing with both the Keyboard and the ImageSphere in the
centre of the screen.

I’ve placed them above all the MenuButtons but I’m not certain that’s
the best place for them either. In the fullness of time they will find
their home, but for now this works fine, and I’m happy to merge this
branch in and consider Search as completed!