First commit into new branch - Added Home button to app!
The app now has 4 buttons in the MenuBar: Home, Search, Profile and
Gallery.

I believe that these will be all the buttons needed in order to build
all the features we’re looking to have for Beta (and possibly beyond).

Buttons within each SubMenu (Profile and Tags for Search; Delete for
Profile; Upload for Gallery) all appear above the main menu bar at the
moment.

I’ll need to create a new component called Home now and have it
communicate with the BackEndAPI and Posts components in order to tap
into the framework we already have for downloading images, and get the
public timeline working!