Selecting a Handle opens up relevant Search box
I quickly added the ability for Handles to call through to Search and
actually open up the profile of the user assigned to that Handle!

I’ve gone beyond the public timeline in fact so I’m going to merge this
branch in and beginning working on Likes in another branch! =)