Selecting the Heart button now impacts the number of likes!
Selecting the HeartButton calls into the ImageSphere, which calls into
Posts, and finally calls through to the BackEndAPI (maybe it should all
be handled in the ImageSphere?), therefore letting the Back-End know
that the user has liked/unliked the image.

Right now I don’t have the knowledge of whether I’ve already liked a
specific post or not, so I’m always beginning with the HeartButton set
as if I don’t like the post. This needs to be fixed from the API’s side.

Next I will make it such that clicking the Like button displays results
for the users who actually like that post…