Created the Likes System based off how Posts works
When selecting the number of likes by a post, it will now call into the
Likes system and display a list of results!

These results are gathered and stored in a very similar way to how I
built the Posts system, and then displayed 10 at a time in front of the
user.

If there are more than 10 likes then a next arrow will appear, and you
can move between results by pressing the next and back arrow buttons.
There’s also a close button, which just closes it all up.

Selecting any user handle will call through to that User in much the
same way as it does when selecting a user’s handle by their post.

I’m checking this in and In principle the Likes system is now finished,
however I’m not going to merge the branch without some more testing for
bugs.