Fixed Back-button and Home-button not working bug + Get AccessToken now on every call regardless of success-status
In order to fix the bug that blocked our submission, I discovered that
I needed to be import Oculus’ “OVR_Unity_Utilities” and add a prefab
that belonged to it into the scene called OVRCameraRig, this has
scripts attached to it that handle a bunch of things including the Back
and Home button presses (in fact I’m amazed I managed to have the
project working at all without Oculus’ utilities now that I’ve
discovered they exist and you need them for every project).

I also changed it so that all BackEndAPI calls will always search for a
new Access-token, as before they would only search for it if there was
a success code, but we might genuinely find the case that a 404 (ie. an
error) is a correct situation (because say User A has deleted a post
that User B is trying to refresh), and this will mean that I will still
get hold of the new Access-Token!