Hooked up ProfileDetails such that it now shows the relevant Following information
ProfileDetails does the work getting all the details and displaying
them appropriately, and the FollowButton is what allows us to follow a
user or not.

Currently to bring up the ProfileDetails you need to press on the
UserMessage when we are displaying the Posts for a user, but this is
not at all obvious and needs to be made clearer to the user!