Activated Personal Timeline + Fixed annoying MenuButton bug!
This means a user can both follow another user and also see a timeline
of the posts from just the users they follow.

For now I’ve also made it such that ProfileDetails and ListUsers cannot
be open at the same time, they were overlaying in a way that had too
much information going on at the same time. However, I may change this
later as when you click on “Followers” it makes the ProfileDetails
disappear, which is also annoying.

I had an annoying MenuButton bug for a long time that was related to
the fact that whenever buttons were clicked and de-activated, then when
re-enabled they would still think the user was gazing at them -
thankfully I fixed this!

The way I managed this was by moving the function hook ups from
OnEnable() and OnDisable() to Start() and OnDestroy() respectively. And
now when OnEnable() is called we simply force the MenuButton to believe
its had not got anyone over it. This change also seems to have removed
a mini frame out that I got when opening the Keyboard, which must have
been related to so many function HookUps that would previously happen
OnEnable() of all the keyboard’s buttons!