Made User Message turn into a button when ProfileDetails are available
Created a MessageButton component that essentially centralises the
responsibility of both setting the text on the central button (as it
was being done haphazardly throughout the code), and allows the bar to
switch between simply displaying text, and being an active button in
itself.

This is really good as it can now tell the user when its a button that
can be pressed to display the ProfileDetails, and also stop being a
button when necessary. And the code to set text on it is infinitely
tidier and simpler throughout.