Added ability to Save picture as a ProfilePic!
Now when opening up a User’s ProfileDetails we see a new ImageSphere
that represents their personal Profile Picture!

In order to save it a user simply picks “Set As Profile Picture”
instead of “Share” when they choose to upload an image.

There are a lot of changes in this commit but they are largely
unimportant, simply ways of me getting the ability to set things
correctly through the ImageSphereController onto a new ImageSphere and
still have it interact appropriately with the ImageSkybox.

I did however make a slight change to the JavaPlugin such that it will
now expect to be given a ThumbnailSize rather than having this value
hardcoded in the plugin, as it gives me more control in the future -
although I haven’t changed the value from its 640 size.