Created new branch called "Comments"
This commit simply sets BackEnd Environment back to Development and
creates a new “Comments” component.

This branch will work on getting commenting functionally finished.
Allowing users to:
(1) View all comments on a post
(2) Edit their comments on a post
(3) Add new comments
(4) Select Hashtags within a comment and be taken to the appropriate
Search tab (this is the most ambitious piece of functionality)