Add Comment functionality created
You can now add a new comment to the list of comments that show up when
you click on a comment. The code and UI follow the same logic as with
the Caption when posting a new picture from the Gallery where a new set
of options come up and you can edit the text appropriately then share
it as a new comment.

I also now prefix every comment with the userhandle of the user who
posted it.

I now need to add (1) the comment count, (2) editing and deletion of
comments (which I think I will do by will simply showing some more UI
when you press on an actual comment).