Deletion and Editing of Comments is now possible!
Two new submenus exist within the Comments allowing the user to Edit +
Delete appropriately.

With this all done we’ve essentially got the entire first version of
the app done in my eyes.

I’m still going to add things within this feature branch in order to
make the experience a bit prettier and smoother, in particular I want
to:
(1) Add a background to every submenu that appears, as it can currently
get quite messy when submenus overlay the ImageSpheres.
(2) Allow hyperlinking of “Handles” and “HashTags” within Comments -
this will be tougher and means that whenever you write a new piece of
text you need to be able to create Physics shapes that fit around the
text, then be able to change its colour at runtime.