Added Backgrounds to all SubMenus!
Backgrounds make it clear that you shouldn’t be pressing on the buttons
behind and stop awkward overlaps between submenu buttons!

I think that in general this had made the whole UI a lot more readable!

I also spotted a problem that when hiding the UI, ImageSphere’s would
still update and show their Metadata, so I’ve added checks within the
ImageSphere’s to whether they should actually be showing the Metadata
or not depending on the MenuController’s state.

TODO: I still need to fix up the ProfileDetails SubMenu as it
intersects with the Keyboard now!