Added a slight floaty Animation to ImageSpheres
This is just a nice touch that I thought I’d add that I can remove at
any time if we decide we don’t like it.

Now ImageSphere’s float a little around their centre of mass, creating
a floaty feeling to them. This is simply done by adding a small
variation every frame to the transform’s x,y,z position based on a sine
wave, this is done in the ImageSphereAnimation component!