Added AnimateSphereTowardsUser() function
In the past it wasn’t entirely clear how the sphere’s represented the
image that you see in the background.

But now when a user selects a sphere it will now come towards him and
he will therefore be able to clearly understand that he’s entering this
new ImageSphere.