Added PlatformMenu component to OVRCamera prefab!
For some reason the prefab which solves all the back button issues,
doesn’t in fact solve all of them without adding another of their
scripts namely “OVRPlatformMenu.cs”.

I added this script and now short presses of the back button is working.

In order for the Oculus Store to accept it I also bumped up the version
code a couple of times to 5 (every time you submit to one of their
build channels, even if its the Alpha one, then the version number must
be upped).

So now we’re on version number: 1.1.1b-2.0.4 and version code: 5!