Added OneSignal to app for PushNotifications!
This was a hell of a lot easier from the client side than I imagined,
OneSignal really makes it super-easy!

Now I just need to get the correct Logo for the PushNotifications (as
I’ve put a temporary one) and get Dan to set it up such that when the
user has (1) new people following him, or (2) new posts from followers,
he will actually receive a notification.

I also need to add the option to disable push-notifications for the app
as that’s also crucial.