Edited ProjectSettings to make building easier for now
COME BACK TO THIS CHECK-IN TO LATER REVERT PROJECT SETTINGS

ProjectSettings now build a package called io.vreel.mox and AndroidManifest allows it to be launched off the phone's main menu (as opposed to through Oculus Store).

Naturally, this will need to be undone when it comes to submitting the app.