Began conversion of Design into a new scene called VReel_v2
The aim here is to convert the Design into a functional scene.

I’ve started with the MenuBar and ImageSphere’s.

Since I don’t have the device I can’t test it, but in principle I
believe that the core elements (all elements on the menu-bar, and the
loading Images themselves) should work - what will not yet work is the
selection of the buttons around the ImageSphere’s.