Push Notifications now work!
With Dan's change from his side I've added the OneSignal-UserID as a parameter to SignIn and CreateUser. Dan then uses these parameters accordingly to send notificaitons when (1) a user has a new follower or (2) any of their followers create a post.

Notifications were interrupting the flow of the app in an very obtrusive way when appearing, but I managed to fix this by simply doing ".InFocusDisplaying(OneSignal.OSInFocusDisplayOption.Notification)" on initialisation.

I also fixed something that's been annoying me for ages, which is that the Keyboard would simply be broken if I made a build with it disabled - but now the AppDirector always enables it on Start() regardless, and it will then hide itself, meaning that I'll never have to do 2 builds again for that reason.