Updated Heart and Like Icons
This took way too long because I needed to fiddle around with all the icons to put them exactly in the right position and scale them appropriately - but the important thing is that I replaced them correctly in the end.

In the process of doing this I also moved a bunch of old Icons into the "Art/Old" folder.