Updated Login to new Design
The entire LoginFlow now uses the new Design - this included updating some prefabs that Mox made and moving them into the Prefabs folder.

The old Design is still in there for reference but should be deleted.

Also added the Reticle back into the DesignPrototyping_04 scene, so it can be tested against the UI mockups.