Login is fully functional with new Design
We were missing a back button in the ForgotPassword menu and the buttons were all switching by default to a dark colour that I had previously set in VRMenuButton.

I also updated the Likes circle, as Mox had given me a new and improved icon!

The only notable issue with the Login is that since we still have the old keyboard, when you select an input field, it will be blocked by the old keyboard - this means we need to (1) update the keyboard, and (2) move the UI up to a better position