Options updated to the new Design
Most of this was just a case of taking the new Design and plastering it over where the old images were - yet its still a time consuming endeavour.

One key difference is that as we hide the ImageSpheres and MenuBar when the Options come up, I have added functionlality to do this in the MenuController.