Updated the Gallery to the new Design
There was the slightly slow process of UI conversion, but the biggest problem I had was making sure that there was no overlapping UI, especially when opening the Options menu.

The solution was simple though, let the MenuController turn on/off the current SubMenu whenever the Options menu comes on/off, according to what state the AppDirector believes we're in.