Updated Profile to follow new Design (mostly)!
The Profile Design update is not entirely finished as I haven't updated the Profile Editing, Follower Listing, or Image Deletion.

However, the primary re-adjustment of the Design has been done!

There was a lot of work here in particular because I began making it such that the ImageSphere on the MenuBar actually displays your own Profile Picture, and that needed to live side by side with the Profile Picture that comes up when a Profile is selected. Both are selectable and can replace the main background image!

Previously I had anybody's profile other than your own be a subpart of Search - but now I'm making it such that it always sets the AppDirector's state to be AppState.kProfile! I've therefore moved all of the ProfileDetails submenu into the ProfileSubMenu in Unity; and code-wise I think that Profile.cs and ProfileDetails.cs should be merged into a single class, as they no longer make sense being separate.

------------------------

Also spotted something very interesting, that Image Loading of images taken with my new Insta360 camera was faster than the 5 seconds I was giving the ImageSphere Animation to come to your face when you select it - this happens because the resolution is far lower with the Insta360 than with the Samsung Gear360. This meant that the high-res image would load in, then be splatted over by the low-res thumbnail!! For now I've fixed this by simply making it take 3 seconds for the image to come to your face - so the thumbnail always get selected before the full image comes in. However, in the long-run I hope to make the loading of the full-image much faster anyway.

I also replaced old SplashScreen and Icon assets with new ones - but this doesn't really make any real difference because these assets are being overwritten by the Oculus Store anyway.