Updated Profile Edit to the new Design!
Profile Editing now follows the new Design. In the process I also added functionality to allow the user to update their username!

Next will be Follower Listing, and Image Deletion.

Also disabled swiping as a form of hiding or unhding the background image as this isn't the way we're going to do it any more and its intefering with my hiding of the spheres behind certain menus.