Updated ListUsers to the new Design
ListUsers now follows the new Design.

This has been an awkward and painful process though, and demonstrates how retrofitting functionality isn't going to work by itself.

Its hard to make parts of the UI appear at certain points and not at other points when you don't know the state that you came from...

Essentially I need to finish off the redesign, then go back and re-do some parts of how the these systems interact, in particular I need to handle many different things based on the state it came from.

I'm checking this in with 2 noticeable problems...
(1) Left/Right arrows are not disappearing correctly in multiple places.
(2) ListUsers is not getting hold of whether one of the entries follows you or not - so it always suggests that you don't follow the user even when you do...