Updated Search to the new Design!
For the most part this works well!

The main things that are yet to be fully updated are the "form input", accompanying keyboard, and maintaining state when selecting a Profile!

Right now if you select a Profile it will totally forget where you are in the Search menu, and pressing back will simply take you to the Explore menu.

I have however fixed the Arrows appearing at incorrect times by introducing a flag in the AppDirector called "OverlayShowing" that when set to true will prohibit the Arrows from displaying - this is then set to true when Options, Profile's Edit Profile, Gallery's Create Post, and ListUsers are open.