Set up most of the new Design for ImageSelection
Created a new class called ImageFlow that controls the switching between summary/comments/options/new-comment.

The only thing that's currently entirely ignored is how to make all the UI disappear.

The Code is also an incredible spider-like mess! This will need to be heavily re-factored and the logic has been: "just get it to work, because then re-factoring will be easier as I will know what classes should be merged and chopped up".

The next order of things will then be to (1) get the keyboard working, (2) refactor the assets [ie. move and remove assets], and (3) refactor the code [this is essential!]...

With that all done then I might be able to move onto other non-Design related things... such as Graphics improvements (which I'm already going to start looking at tomorrow with Andy Thomason), and porting the Analytics over to Mixpanel or Google Analytics.