Updated Keyboard to new Design!
This was relatively quick because Mox had done most of the work already! The only thing I have yet to do (other than the animations of course), is to make the "Type" box work. This is still a crucial improvement to the keyboard that needs to be done at some point soon, because it will make typing a lot easier, as atm its a real hassle to look down to type, then up to see what you've typed!