Created MenuHider
This is essentially a Collider that lives in the ImageFlow submenu, and when the user looks at it, it will hide all the Menu UI and undim the background image!

If the user then clicks on this hidden collider box (which shouldn't be too hard because its right in the centre where the UI normally appears), then it will bring the UI back and dim the background again.

The sensation is that when you've selected an Image and you've looked away from the UI, then it will hide it all, and when you want you can bring it back!