Asset Tidy-up
Moved all assets that are being used in the main scene out of MoxDesign and into their respective top-level folders.

I've tried to remove use of duplicate assets, and have just the one in my main folders too.

Essentially this is the first step in getting the project, and main scene into a much tidier state.