VReel scene tidy!
Deleted lots of Old and unnecessary gameobjects/components, the scene in fact looks cleaner than it ever has in my opinion (!), although there are obviously many bugs that I still need to work through. The only part of the scene that I haven't cleaned up are the ImageSphere, which I will do next!