Tidied ImageSphere's
The tidying part was actually very fast, and I just applied it to the prefab.

However, I lost a lot of time stupidly chasing a non-existent bug, as I couldn't work out why the Small Profile sphere's weren't loading... it turned out the reason was simply because I'm in Staging, and my profile didn't have a ProfilePic in staging!