Code Tidy
Removed any mentions of UserMessage as it doesn't exist anymore in the new Design - although this still leaves us with a few cases where the user does not get any proper feedback on his actions.

Merged the Profile Functions into ProfileDetails (that I still want to rename to become Profile again, as I think that's the more correct name!).

This is all totally WIP and there's still so much code to refactor and bugs to fix!