Reworked MenuController's Visibility activation
MenuController now has a list of MenuConfigs that any individual class can access and create its own of through RegisterToUseMenuConfig().

By updating this config, then calling MenuController.UpdateMenuConfig(), the MenuController will switch on/off the relevant parts of the menu correctly, however it will not overwrite settings that another class may have stored on its own MenuConfig.

What's great about this is that each class can mess around with the visible elements of the UI as much as it wants, without the previous state being forgotten, because there are literally many different states.

In terms of functionality it appears like nothings changed, as the app should behave exactly the same - however, in practice this is a massive improvement to the code and its stability in the long-run!