Imported MixPanel + Improved Anti-Aliasing on Logo (through softening the edges of the texture)
Imported MixPanel's UnityPackage since I'm going to begin using it for our Analytics, instead of using AWS's version which is not particularly good.

Also, now that most of the button images are at a good resolution, the main problem that's visible is alisasing. Since I wasn't able to add anti-aliasing when working with Andy Thomason, I tried to look into how to do this with the textures itself.

So I asked Mox to add a feather to the Logo Icon (which was an obvious example of too much aliasing), and this has massively reduced the aliasing - I think going forward we'll need to add a feather to soften the edges of all images in order to essentially do a form of anti-aliasing within the textures themselves!