WIP Analytics (Identify/Alias not working)
For the most part the Analytics tracking is working. However, I'm having a frustrating issue with Mixpanel where I can't seem to get Logging out to work correctly and all users are coming in as the same user - I've got the Indetify(), Alias() and Reset() calls interspersed in the code at where I believe they should be, but it doesn't seem to be working correctly.

I'm checking this in, in a WIP state so that I can carry on with other things, but I will send Mixpanel a request to see if I can have their support help me out.

I'm also checking in a bunch of Logo images, as I'm trying to work out how best to remove scintillation from all the images with Mox, starting with the Logo.