Updated Analytics Text, and ImageSphere ImageSelected()
I read that its best to give more semantic meaning to Analytics, so I've changed fields like "FollowSelected" to "Follow Tab Openend"! I did this for every button and ensure that they are all getting correctly called.

I also spotted that ImageSphere selection is following an old paradigm where I would manually hook up to the MenuButton component. I've since changed it such that MenuButtons are only triggered by function hook ups in the scene because its easier to figure out the functions getting called) - although its fine if this rule isn't obeyed 100%, I still found it better to update it for ImageSelection (which is the mos tcentrla concept of the app).