Updated the Logo removing Scintillation (other images to follow)
The main problem here was that the edges were not soft enough.

Therefore, when we had high-resolution (eg. with mipmaps turned off and a 720x720 image), it would cause scintillation due to the high frequency sampling on the edges.

On the flip side, when we had low-resolution (because say mipmaps were switched on), then the scintillation at the edges would go away but the image would just look like it was at a low quality.

Mox has fixed this by: (1) feathering the edge; (2) adding a 1px blur on entire object - and giving me the logo in 310x310 resolution. Due to an unfortunate situation, if I switched on mipmapping on the logo then the current distance of the camera means that it would drop to a lower mip and the lower resolution (I assume 155x155) looks a lot worse, so for now I've just switched off mipmapping on the logo, but the correct solution would be to have a shader where we can change the mip-bias.