Created Guest Account Access!
This was easier than I thought it might be!

Essentially all that I had to do was remove Login from being an AppState within the AppDirector, and begin treating it just as an Overlay like the Options Menu.

I then put checks throughout the code for whenever the user needs to be logged in to do an action, it will now open the LoginSubMenu instead of doing that action (eg. whenever the user clicks Guest on the MenuBar, or Heart on an Image, then it will open the LoginFlow instead of performing that action).